feat: [FC-0063] Application is dockerized
