+++ .dockerignore
@@ -1,0 +1,7 @@
+.git
+.github
+
+**/.coverage
+**/.pytest_cache
+**/__pycache__
+**/*.pyc
