+++ setup.py
@@ -1,26 +1,12 @@
-import os
-import re
-
 from glob import glob
 from os.path import basename, splitext
 
 from setuptools import setup, find_packages
 
+from utils import get_version, load_requirements
+
 with open("README.rst", encoding="utf-8") as readme_file:
     readme = readme_file.read()
-
-
-def get_version(*file_paths):
-    """
-    Extract the version string from the file at the given relative path fragments.
-    """
-    filename = os.path.join(os.path.dirname(__file__), *file_paths)
-    version_file = open(filename, encoding="utf-8").read()
-    version_match = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]", version_file, re.M)
-    if version_match:
-        return version_match.group(1)
-    raise RuntimeError("Unable to find version string. ")
-
 
 VERSION = get_version("src", "cc2olx", "__init__.py")
 
@@ -41,7 +27,7 @@ setup(
     ],
     description=("Command line tool, that converts Common Cartridge " "courses to Open edX Studio imports."),
     entry_points={"console_scripts": ["cc2olx=cc2olx.main:main"]},
-    install_requires=["lxml"],
+    install_requires=load_requirements("requirements/base.txt"),
     license="GNU Affero General Public License",
     long_description=readme,
     include_package_data=True,
